secondary_server/api_service/requirements.txt
psycopg2-binary==2.9.3
requests==2.26.0